server/requirements.txt
# Core Django
Django==4.2
asgiref==3.8.1
sqlparse==0.4.4
pytz==2022.7.1
tzdata==2024.1
backports.zoneinfo==0.2.1;python_version<"3.9"

# Django REST Framework
djangorestframework==3.14.0
djangorestframework-simplejwt==4.7.1
drf-yasg==1.21.7
drf-nested-routers==0.93.5
django-filter==23.5

# Django Extensions
django-cors-headers==3.7.0
whitenoise==6.6.0

# Database
psycopg2-binary==2.9.9

# Authentication
PyJWT==2.1.0

# Environment & Configuration
python-dotenv==1.0.1

# Production Server
gunicorn==20.1.0

# Core Dependencies
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
urllib3==2.2.1
requests==2.31.0

# DRF-YASG Dependencies
uritemplate==4.1.1
inflection==0.5.1
PyYAML==6.0.1
MarkupSafe==2.1.5
Jinja2==3.1.4

# Common Dependencies
packaging==24.0
typing_extensions==4.11.0
setuptools<81.0.0  # Pin to avoid pkg_resources deprecation warning from drf-yasg
